♻️ run_service self-improvement loop 빌더 모듈 추출 (split_now 해소)
변경 이유
- WT2 wiring 으로 run_service.py 가 1016 LOC 가 되어 code_audit split_now(>1000+책임 다수) 위반. self_improvement 책임이 audit 에서 분리 신호로도 표면됨

주요 변경 사항
- runtime/supervisor_self_improvement.build_self_improvement_loop 로 빌더 추출(886 LOC 로 감소)
- run_service supervisor 분기는 import 해서 호출만

비고
- 동작/플래그 동일 — 순수 구조 분리. code_audit split_now 0 복구
